@@ -7,38 +7,55 @@
     "dev": "next dev",
     "build": "NODE_ENV=production next build",
     "start": "next start",
-    "export": "next export",
-    "export:hands-on-workshop": "GORIGHT_EXPORT=/hands-on-workshop next export -o hands-on-workshop/",
+    "lint": "next lint .",
     "deploy": "gh-pages -d ./out"
   },
   "dependencies": {
-    "@bridge-the-gap/design-system": "2.4.2",
-    "@svgr/webpack": "^5.5.0",
-    "@tailwindcss/aspect-ratio": "^0.2.1",
-    "copy-webpack-plugin": "^10.0.1",
-    "date-fns": "^2.30.0",
+    "@bridge-the-gap/design-system": "2.5.0",
+    "@emotion/react": "^11.11.4",
+    "@mdx-js/loader": "^3.0.1",
+    "@mdx-js/react": "^3.0.1",
+    "@next/mdx": "^14.2.3",
+    "@svgr/webpack": "^8.1.0",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.13",
+    "copy-webpack-plugin": "^12.0.2",
+    "date-fns": "^3.6.0",
     "gray-matter": "^4.0.3",
     "html-to-image": "^1.11.11",
-    "next": "12",
-    "next-seo": "^4.29.0",
-    "react-calendly": "^4.3.0",
+    "imgix": "^0.0.2",
+    "lodash": "^4.17.21",
+    "next": "14.2.3",
+    "next-seo": "^6.5.0",
+    "react-calendly": "^4.3.1",
     "react-ga": "^3.3.1",
-    "string.prototype.replaceall": "^1.0.7",
-    "stringify-object": "^3.3.0",
-    "theme-ui": "0.6.1"
+    "remark-embed-images": "^4.0.0",
+    "remark-frontmatter": "^5.0.0",
+    "remark-images": "^4.0.0",
+    "remark-mdx-frontmatter": "^4.0.0",
+    "remark-mdx-images": "^3.0.0",
+    "remark-unwrap-images": "^4.0.0",
+    "string.prototype.replaceall": "^1.0.10",
+    "stringify-object": "^5.0.0",
+    "theme-ui": "0.16.2"
   },
   "devDependencies": {
-    "@mdx-js/loader": "^1.6.22",
-    "@next/mdx": "^12.3.4",
-    "autoprefixer": "^10.4.14",
-    "postcss": "^8.4.24",
-    "react": "17",
-    "react-dom": "17",
-    "tailwindcss": "^3.3.2"
+    "autoprefixer": "^10.4.19",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.3",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-simple-import-sort": "^12.1.0",
+    "next-export-optimize-images": "^4.4.0",
+    "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.6.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.3"
   },
   "peerDependencies": {
-    "react": "17",
-    "react-dom": "17"
+    "react": "18",
+    "react-dom": "18"
   },
   "packageManager": "yarn@1.22.19"
 }
