@@ -1,9 +1,9 @@
 /* simplified version of this article: https://www.josephrex.me/frontmatter-with-nextjs-and-mdx/ */
-const matter = require("gray-matter");
-const stringifyObject = require("stringify-object");
+import matter from "gray-matter";
+import stringifyObject from "stringify-object";
 
-module.exports = () => (tree, file) => {
-  const { data } = matter(file.contents);
+const frontmatterPlugin = () => (tree, file) => {
+  const { data } = matter(file.value);
   const { layout, ...props } = data;
 
   // Remove frontmatter after converting it into JS object
@@ -21,14 +21,12 @@ module.exports = () => (tree, file) => {
   tree.children.unshift(
     {
       type: "import",
-      value: `
-      import { ${layout} } from '@components/layout'
-    `,
+      value: `import { ${layout} } from '@components/layout;'`,
     },
     {
       type: "jsx",
       value: `
-    <${layout} {...${stringifyObject(props)}}>
+      <${layout} {...${stringifyObject(props)}}>
     `,
     }
   );
@@ -36,16 +34,14 @@ module.exports = () => (tree, file) => {
   // Close JSX parent component
   tree.children.push({
     type: "jsx",
-    value: `
-    </${layout}>
-    `,
+    value: `</${layout}>`,
   });
 
   // Convert frontmatter to JS object and push to document tree
   tree.children.push({
     type: "export",
-    value: `
-    export const meta = ${stringifyObject(props)}
-    `,
+    value: `export const meta = ${stringifyObject(props)}`,
   });
 };
+
+export default frontmatterPlugin;
