@@ -1,21 +1,22 @@
-import { toPng } from 'html-to-image';
-import format from 'date-fns/format';
+import format from "date-fns/format";
+import { toPng } from "html-to-image";
 
 export const downloadPng = (ref) => {
-    const getFileName = (fileType) => `${format(new Date(), "'PartsOfDesignSystem-'HH-mm-ss")}.${fileType}`;
+  const getFileName = (fileType) =>
+    `${format(new Date(), "'PartsOfDesignSystem-'HH-mm-ss")}.${fileType}`;
 
-    if (ref.current === null) {
-        return;
-    }
+  if (ref.current === null) {
+    return;
+  }
 
-    toPng(ref.current, { cacheBust: true, backgroundColor: 'white' })
-        .then((dataUrl) => {
-            const link = document.createElement('a');
-            link.download = getFileName('png');
-            link.href = dataUrl;
-            link.click();
-        })
-        .catch((err) => {
-            console.log(err);
-        });
+  toPng(ref.current, { cacheBust: true, backgroundColor: "white" })
+    .then((dataUrl) => {
+      const link = document.createElement("a");
+      link.download = getFileName("png");
+      link.href = dataUrl;
+      link.click();
+    })
+    .catch((err) => {
+      console.log(err);
+    });
 };
