@@ -1,5 +1,5 @@
 export { default as LayoutDefault } from "./layout";
-export { default as LayoutHandout } from "./layoutHandout";
-export { default as LayoutWorkshop } from "./layoutWorkshop";
 export { default as LayoutBlog } from "./layoutBlog";
 export { default as LayoutCase } from "./layoutCase";
+export { default as LayoutHandout } from "./layoutHandout";
+export { default as LayoutWorkshop } from "./layoutWorkshop";
