@@ -1,5 +1,5 @@
-import Header from "@components/header";
 import Footer from "@components/footer";
+import Header from "@components/header";
 
 const navItems = [
   {
@@ -28,9 +28,7 @@ export default function Layout({ children }) {
   return (
     <div className="container flex flex-col justify-between min-h-screen overflow-hidden pt-11">
       <Header navItems={navItems} />
-      <main className="max-w-full text-light-on-background-900">
-        {children}
-      </main>
+      <main className="max-w-full text-light-on-background-900">{children}</main>
       <Footer />
     </div>
   );
