@@ -1,23 +1,13 @@
-import { MDXProvider } from "@mdx-js/react";
+import { CtaLink, Text } from "@bridge-the-gap/design-system";
 import Link from "@components/link";
-import { Text, CtaLink } from "@bridge-the-gap/design-system";
+import MDXComponents from "@components/MDXComponents";
+import { MDXProvider } from "@mdx-js/react";
 import { useRouter } from "next/router";
 import { NextSeo } from "next-seo";
 
-
-const MDXComponents = {
-  a: Link,
-};
-export default function HandoutLayout({
-  title,
-  seo,
-  layout,
-  children,
-  ...props
-}) {
+export default function LayoutHandout({ title, seo, ...props }) {
   const router = useRouter();
   const currentPath = router.asPath;
-  const meta = { title, ...seo };
 
   const getParentPath = (currentPath) => {
     if (currentPath.endsWith("/")) {
@@ -48,24 +38,16 @@ export default function HandoutLayout({
       <div className="grow py-8 bg-white border-b">
         <div className="flex bg-white min-w-100">
           <div className="container max-w-2xl py-2 mx-auto mb-8 text-xl text-left">
-            <CtaLink
-              arrow="start"
-              as={Link}
-              href={mainPagePath}
-              underline={true}
-            >
+            <CtaLink arrow="start" as={Link} href={mainPagePath} underline={true}>
               Back to the Main page
             </CtaLink>
           </div>
         </div>
-        <Text
-          variant="6Xl"
-          className="w-full my-2 text-center text-light-on-background-900"
-        >
+        <Text variant="6Xl" className="w-full my-2 text-center text-light-on-background-900">
           {title}
         </Text>
         <div className="px-4 mx-auto my-8 prose text-justify ">
-          <MDXProvider components={MDXComponents}>{children}</MDXProvider>
+          <MDXProvider components={MDXComponents}>{props.children}</MDXProvider>
         </div>
       </div>
     </>
