@@ -1,6 +1,6 @@
 import fs from "fs";
-import path from "path";
 import matter from "gray-matter";
+import path from "path";
 
 const postsDirectory = path.join(process.cwd(), "pages/blog");
 const casesDirectory = path.join(process.cwd(), "pages/case-studies");
@@ -9,10 +9,7 @@ export function getPostSlugs() {
   let postSlugs = [];
   fs.readdirSync(postsDirectory).forEach((file) => {
     const currDirectoryPath = path.join(postsDirectory, file);
-    if (
-      fs.existsSync(currDirectoryPath) &&
-      fs.lstatSync(currDirectoryPath).isDirectory()
-    ) {
+    if (fs.existsSync(currDirectoryPath) && fs.lstatSync(currDirectoryPath).isDirectory()) {
       postSlugs.push(path.basename(currDirectoryPath));
     }
   });
@@ -58,10 +55,7 @@ export function getCasesSlugs() {
   let casesSlugs = [];
   fs.readdirSync(casesDirectory).forEach((file) => {
     const currDirectoryPath = path.join(casesDirectory, file);
-    if (
-      fs.existsSync(currDirectoryPath) &&
-      fs.lstatSync(currDirectoryPath).isDirectory()
-    ) {
+    if (fs.existsSync(currDirectoryPath) && fs.lstatSync(currDirectoryPath).isDirectory()) {
       casesSlugs.push(path.basename(currDirectoryPath));
     }
   });
@@ -88,8 +82,6 @@ export function getThumbnailPath(slug) {
     console.log("directory doesn't exist: ", srcPath);
     return null;
   }
-  const thumb = fs
-    .readdirSync(srcPath)
-    .find((filename) => filename.startsWith("thumb."));
+  const thumb = fs.readdirSync(srcPath).find((filename) => filename.startsWith("thumb."));
   return thumb ? slug + "/" + thumb : null;
 }
