@@ -1,9 +1,9 @@
 const replaceAll = require("string.prototype.replaceall");
 
-module.exports = function (content, map, meta) {
+module.exports = function (content) {
   return replaceAll(
     content,
     /\!\[(.*)\]\((\.\/.+)\)/g, // caprtures only images with relative src, like "![](./zoom-breakout-rooms2.png)""
-    `<Image alt="$1" src={require('$2').default} />`
+    `<img alt="$2" src={require('$2').default} />`
   );
 };
